+++ packages/gu-ds-css/CHANGELOG.md
@@ -1,5 +1,13 @@
 # gu-ds-css
 
+## 0.0.3
+
+### Patch Changes
+
+- Added progress bar components
+- Updated dependencies
+  - gu-ds-tokens@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes