+++ packages/gu-ds-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # gu-ds-tokens
 
+## 0.0.3
+
+### Patch Changes
+
+- Added progress bar components
+
 ## 0.0.2
 
 ### Patch Changes